Move PartnersBlock to service app
