+++ tbx/project_styleguide/templates/patterns/molecules/streamfield/blocks/partners_block.html
@@ -7,6 +7,6 @@
 
 <div class="partners__image-card">
     {% for logo in value.partner_logos %}
-        {% image logo.image fill-100x100 format-webp alt="" %}
+        {% image logo fill-100x100 format-webp alt="" %}
     {% endfor %}
 </div>
