+++ tbx/project_styleguide/templates/patterns/molecules/streamfield/blocks/partners_block.yaml
@@ -9,6 +9,6 @@ context:
 
 tags:
   image:
-    logo.image fill-100x100 format-webp alt="":
+    logo fill-100x100 format-webp alt="":
       raw: |
         <img src="https://source.unsplash.com/100x100?logo" width="100" height="100" alt="" class="partners__logo">
