+++ tbx/services/blocks.py
@@ -1,6 +1,7 @@
 from tbx.core.blocks import StoryBlock
 from wagtail import blocks
 from wagtail.blocks import CharBlock, StructBlock
+from wagtail.images.blocks import ImageChooserBlock
 
 
 class ShowcaseBlock(StructBlock):
@@ -25,8 +26,25 @@ class ShowcaseBlock(StructBlock):
         template = "patterns/molecules/streamfield/blocks/showcase_block.html"
 
 
+class PartnersBlock(blocks.StructBlock):
+    title = blocks.CharBlock(
+        max_length=255,
+    )
+    partner_logos = blocks.ListBlock(ImageChooserBlock(), label="Logos")
+
+    class Meta:
+        icon = "openquote"
+        label = "Partner logos"
+        template = "patterns/molecules/streamfield/blocks/partners_block.html"
+
+
 class ServiceStoryBlock(StoryBlock):
-    # Dev note - include partners_block here when merged,
+    # Dev note - include work listing and blog listing blocks here when merged
+    partners_block = PartnersBlock(
+        icon="openquote",
+        label="Partners logos",
+        template="patterns/molecules/streamfield/blocks/partners_block.html",
+    )
     showcase = ShowcaseBlock(
         icon="tasks",
         template="patterns/molecules/streamfield/blocks/showcase_block.html",
